Building controller, service, repository, config classes.

+++ src/main/java/lgrimm1/Books/Series/SeriesService.java
@@ -1,7 +1,7 @@
 package lgrimm1.Books.Series;
 
+import lgrimm1.Books.Books.*;
 import org.springframework.beans.factory.annotation.*;
-import org.springframework.http.*;
 import org.springframework.stereotype.*;
 
 import java.util.*;
@@ -10,40 +10,59 @@ import java.util.*;
 public class SeriesService {
 
 	private final SeriesRepository seriesRepository;
+	private final BookRepository bookRepository;
 
 	@Autowired
-	public SeriesService(SeriesRepository seriesRepository) {
+	public SeriesService(SeriesRepository seriesRepository, BookRepository bookRepository) {
 		this.seriesRepository = seriesRepository;
+		this.bookRepository = bookRepository;
 	}
 
-	public SeriesEntity create(SeriesEntity seriesEntity) {
-		return seriesRepository.save(seriesEntity);
+	public SeriesEntity createNewEntity(SeriesEntity newSeriesEntity) {
+		if (seriesRepository.findByTitle(newSeriesEntity.getTitle()).isPresent()) {
+			return null;
+		}
+		return seriesRepository.save(newSeriesEntity);
 	}
 
-	public List<SeriesEntity> getAll() {
+	public List<SeriesEntity> getAllEntities() {
 		return seriesRepository.findAll();
 	}
 
-	public SeriesEntity getById(long id) {
+	public SeriesEntity getEntityById(long id) {
 		 Optional<SeriesEntity> optionalSeriesEntity = seriesRepository.findById(id);
 		 return optionalSeriesEntity.orElse(null);
 	}
 
-	public SeriesEntity update(SeriesEntity newSeriesEntity) {
-		long id = newSeriesEntity.getId();
+	public SeriesEntity updateEntity(SeriesEntity modifiedSeriesEntity) {
+		long id = modifiedSeriesEntity.getId();
 		if (id == 0) {
 			return null;
 		}
-		if (seriesRepository.findById(id).isEmpty()) {
+		if (!seriesRepository.existsById(id)) {
 			return null;
 		}
-		SeriesEntity oldSeriesEntity = seriesRepository.getReferenceById(id);
-		oldSeriesEntity.setTitle(newSeriesEntity.getTitle());
-		oldSeriesEntity.setRemarks(newSeriesEntity.getRemarks());
-		return oldSeriesEntity;
+		Optional<SeriesEntity> seriesEntityWithIdenticalTitle = seriesRepository.findByTitle(modifiedSeriesEntity.getTitle());
+		if (seriesEntityWithIdenticalTitle.isPresent() && seriesEntityWithIdenticalTitle.get().getId() != id) {
+			return null;
+		}
+		SeriesEntity originalSeriesEntity = seriesRepository.getReferenceById(id);
+		originalSeriesEntity.setTitle(modifiedSeriesEntity.getTitle());
+		originalSeriesEntity.setRemarks(modifiedSeriesEntity.getRemarks());
+		return originalSeriesEntity;
 	}
 
-	public void delete(long id) {
+	public boolean deleteEntityById(long id) {
+		if (id == 0) {
+			return false;
+		}
+		if (!seriesRepository.existsById(id)) {
+			return false;
+		}
+		if (bookRepository.findBySeriesId(id).isPresent()) {
+			return false;
+		}
 		seriesRepository.deleteById(id);
+		return true;
 	}
 }
